♻️ Merge duration.go to ops/duration.go
